Updated loader component
