+++ sources/packages/web/src/composables/useFormioDropdownLoader.ts
@@ -67,30 +67,6 @@ export function useFormioDropdownLoader() {
     );
   };
 
-  // Get Program description for the selected program and set to the field (programDesc) in formio
-  const loadProgramDesc = async (
-    form: any,
-    programId: number,
-    fieldId: string,
-  ) => {
-    const valueToBeLoaded = await EducationProgramService.shared.getStudentEducationProgram(
-      programId,
-    );
-    formioUtils.setComponentValue(
-      form,
-      fieldId,
-      `<h3><span style="background-color:transparent;color:#2965c5;"><strong>${valueToBeLoaded.name}</strong></span></h3>${valueToBeLoaded.description}
-      <table style="width:100%">
-      <tr>
-      <th>Credential</th><th>Delivery Method</th>
-      </tr>
-      <tr>
-      <td>${valueToBeLoaded.credentialType}</td><td>${valueToBeLoaded.deliveryMethod}</td>
-      </tr>
-      </table>`,
-    );
-  };
-
   // Retrieve the list of offerings for a particular location.
   const loadOfferingsForLocation = async (
     form: any,
@@ -105,22 +81,6 @@ export function useFormioDropdownLoader() {
         locationId,
         programId,
       ),
-    );
-  };
-
-  // get offering date of the selected offering and set to the hidden field (selectedOfferingDate) in formio
-  const loadSelectedOfferingDate = async (
-    form: any,
-    offeringId: number,
-    fieldId: string,
-  ) => {
-    const valueToBeLoaded = await EducationProgramOfferingService.shared.getProgramOfferingDetails(
-      offeringId,
-    );
-    formioUtils.setComponentValue(
-      form,
-      fieldId,
-      valueToBeLoaded?.studyStartDate,
     );
   };
 
@@ -162,10 +122,8 @@ export function useFormioDropdownLoader() {
     loadProgramsForLocation,
     loadProgramsForLocationForInstitution,
     loadOfferingsForLocation,
-    loadSelectedOfferingDate,
     loadOfferingsForLocationForInstitution,
     loadInstitutionTypes,
     loadPIRDeniedReasonList,
-    loadProgramDesc,
   };
 }
